requirements.txt
argparse
boto3
termcolor
backoff
jira
packaging
GitPython
